+++ R/MARSS-methods.R
@@ -8,22 +8,6 @@ MARSSfit.default <- function(x, ...){
 
 MARSSfit.kem <- function(x, ...) MARSSkem(x)
 
-MARSSfit.EM_KFAS <- function(x, ...){
-  MARSSkem(x)
-}
+MARSSfit.BFGS <- function(x, ...) MARSSoptim(x)
 
-MARSSfit.EM_KFSS <- function(x, ...){
-  MARSSkem(x)
-}
 
-MARSSfit.BFGS <- function(x, ...){
-  MARSSoptim(x)
-}
-
-MARSSfit.BFGS_KFAS <- function(x, ...){
-  MARSSoptim(x)
-}
-
-MARSSfit.BFGS_KFSS <- function(x, ...){
-  MARSSoptim(x)
-}